requirements.txt
llama-index
sqlalchemy
ipython
plotly
python-dotenv
streamlit
pandas
numpy
openai